Your Name: Matt L
Subject Line: Adds html2canvas with mixed results

What Happened: I'm able to take the chart and open it in a new window, which is sweet.... except for the fact that the image
is entirely screwed up. I think that this library will be the solution ultimately, it just needs some attention.