Your Name: Matt
Subject Line: Got a PNG generated from the SVG

What Happened: Right now, the image is appending to the body. We need to change it to open up a download prompt. There's some extraneous crap from a library I tried to use, but I thought you might want to see the progress.