Your Name: Matt
Subject Line: csv

What Happened: Not sure.

I don't think CSV formats are quite strong enough to be reliable.

I am having trouble getting the data to normalize for Google charts and have been stuck with this nasty "not an array" error from that api for awhile.

I duplicated some code in this but I thought it was better for testing before refactoring.

Check it out.