Your Name: Matt Subject Line: Please see the commit before this for most changes What Happened: This commit was simply to add a longer note.
There was a problem with the .dcModal css that I moved into the desktop view. I also added some javascript to change the dimensions of the PNG modal. I am not sure if you'd have a more elegant solution, but it's all working now.

Bad news about the button color on exit.

It appears that is part of the accessibility options that bootstrap provides. I say we just ditch their code and go with the fix. Or we can try the css selector method you had mentioned in the last comment.

Here are the results from my Bug Report if you'd like to see:
https://github.com/twbs/bootstrap/issues/14634