Your Name: matt
Subject Line: linked in auth for dropchart

What Happened:

took a few hours, we are now getting the linked in info properly.

i'm a bonehead and didn't realize .result() was asynchronous.

i'm leaving a place in that result to send the info to the console, still in the air a bit about how to save the user info and charts.

let's talk front-end frameworks when you get back, it needs something.

- ml